refactor: :poop: merge types into one file
