+++ package.json
@@ -3,7 +3,7 @@
   "version": "1.1.5",
   "description": "The validatorjs library makes data validation in JavaScript very easy in both the browser and Node.js.",
   "main": "dist/main.js",
-  "types": "typings/main.d.ts",
+  "types": "dist/main.d.ts",
   "scripts": {
     "dev": "tsc -w",
     "clean": "rimraf dist",
@@ -24,8 +24,7 @@
     "url": "git+https://github.com/chantouchsek/validatorjs.git"
   },
   "files": [
-    "dist",
-    "typings"
+    "dist"
   ],
   "keywords": [
     "validator",
