--- src/types/validator.ts
@@ -1,9 +1,0 @@
-import type { LangTypes } from './lang'
-
-export interface ValidatorOptions {
-  locale?: LangTypes
-  confirmedReverse?: boolean
-  customMessages?: Record<string, any>
-  customAttributes?: Record<string, any>
-}
-export type VoidFunction = boolean | ((...arg: any) => any)
